Moved jwt auth class to separate file to avoid hard dependency.
