+++ dj_rest_auth/utils.py
@@ -1,5 +1,7 @@
 from importlib import import_module
 
+from django.conf import settings
+
 
 def import_callable(path_or_callable):
     if hasattr(path_or_callable, '__call__'):
@@ -15,38 +17,14 @@ def default_create_token(token_model, user, serializer):
     return token
 
 
-try:
-    from django.conf import settings
-    from rest_framework_simplejwt.authentication import JWTAuthentication
+def jwt_encode(user):
     from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
+    TOPS = import_callable(getattr(settings, 'JWT_TOKEN_CLAIMS_SERIALIZER', TokenObtainPairSerializer))
+    refresh = TOPS.get_token(user)
+    return refresh.access_token, refresh
 
-    def jwt_encode(user):
-        TOPS = import_callable(getattr(settings, 'JWT_TOKEN_CLAIMS_SERIALIZER', TokenObtainPairSerializer))
-        refresh = TOPS.get_token(user)
-        return refresh.access_token, refresh
 
-    class JWTCookieAuthentication(JWTAuthentication):
-        """
-        An authentication plugin that hopefully authenticates requests through a JSON web
-        token provided in a request cookie (and through the header as normal, with a
-        preference to the header).
-        """
-        def authenticate(self, request):
-            cookie_name = getattr(settings, 'JWT_AUTH_COOKIE', None)
-            header = self.get_header(request)
-            if header is None:
-                if cookie_name:
-                    raw_token = request.COOKIES.get(cookie_name)
-                else:
-                    return None
-            else:
-                raw_token = self.get_raw_token(header)
-
-            if raw_token is None:
-                return None
-
-            validated_token = self.get_validated_token(raw_token)
-            return self.get_user(validated_token), validated_token
-
+try:
+    from .jwt_auth import JWTCookieAuthentication
 except ImportError:
-    raise ImportError("rest-framework-simplejwt needs to be installed")
+    pass
